Refactor: Move API imports to services module and implement error handling for YouTube Music API
