--- spotifysaver/apis/__init__.py
@@ -1,9 +1,0 @@
-from spotifysaver.apis.spotify_api import SpotifyAPI
-from spotifysaver.apis.youtube_api import YoutubeMusicSearcher
-from spotifysaver.apis.lrclib_api import LrclibAPI
-
-__all__ = [
-    "SpotifyAPI",
-    "YoutubeMusicSearcher",
-    "LrclibAPI"
-]
